# src/app/backend/resources/swagger/cryptocurrency-open-api.json
{
  "openapi": "3.0.0",
  "info": {
    "version": "1.0.3",
    "title": "CNK Cryptocurrency",
    "description": "API to manage CNK cryptocurrency",
    "contact": {
      "name": "Juan Sebastián Bravo",
      "email": "jsbravoc@gmail.com",
      "url": "https://jsbravoc.com"
    },
    "license": {
      "name": "BSD-3-Clause",
      "url": "https://opensource.org/licenses/BSD-3-Clause"
    }
  },
  "servers": [
    {
      "url": "http://localhost:3000"
    }
  ],
...
  "paths": {
...
    "/cryptocurrency": {
      "get": {
        "tags": ["cryptocurrency"],
        "summary": "Retrieves all transactions registered in the blockchain.",
        "description": "Retrieves all transactions registered in the blockchain. Note that, by default, Hyperledger Sawtooth REST API will return up to 100 transactions.",
        "operationId": "getCryptocurrency",
        "parameters": [
          {
            "name": "expanded",
            "in": "query",
            "description": "Expand recursively the transactions",
            "required": false,
            "schema": {
              "type": "boolean",
              "format": "boolean",
              "example": true
            }
          },
          {
            "name": "hidePending",
            "in": "query",
            "description": "Hide pending transactions",
            "required": false,
            "schema": {
              "type": "boolean",
              "format": "boolean",
              "example": true
            }
          },
          {
            "name": "hideInvalid",
            "in": "query",
            "description": "Hide invalid transactions",
            "required": false,
            "schema": {
              "type": "boolean",
              "format": "boolean",
              "example": true
            }
          },
          {
            "name": "limit",
            "in": "query",
            "description": "Limit queried transactions",
            "required": false,
            "schema": {
              "type": "integer",
              "format": "int32",
              "example": 100
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Information about the cryptocurrency flow.",
            "content": {
              "application/json": {
                "examples": {
                  "Transaction array": {
                    "$ref": "#/examples/NonExpandedTransactionArray"
                  },
                  "Expanded transaction array": {
                    "$ref": "#/examples/ExpandedTransactionArray"
                  }
                }
              }
            }
          },
          "503": {
            "description": "Sawtooth service unavailable",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/definitions/Sawtooth503Error"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": ["cryptocurrency"],
        "summary": "Creates a new transaction",
        "description": "Allows the creation of a new transaction, which can be a coinbase (no-input) transaction or a normal transaction. A coinbase transaction must not have any sender defined and also its recipient must have **coinbase** permission set to _true_. This endpoint validates balance and also existing & active users, thus `sender` must have the specified balance and both the `sender` and the `recipient` must exist and be active. **Note:** When ENV variable `DISABLE_INTEGRITY_VALIDATION` is set to _false_, the signature of the transaction must be created with the user's private key (thus, decryptable with their public_key) and its contents must match the transactions sensitive info (`sender`, `recipient`, `amount`, `creator`).",
        "operationId": "createTransaction",
        "requestBody": {
          "description": "New transaction to add to the blockchain",
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/definitions/SimpleTransaction"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Transaction created successfully",
            "content": {
              "application/json": {
                "examples": {
                  "Transaction with supporting_transactions": {
                    "$ref": "#/examples/CreatedTransaction"
                  },
                  "Coinbase transaction": {
                    "$ref": "#/examples/CreatedCoinbaseTransaction"
                  }
                }
              }
            },
            "links": {
              "GetTransactionByAddress": {
                "operationId": "getTransactionByAddress",
                "parameters": {
                  "address": "$response.body#/payload/signature"
                },
                "description": "The `signature` value returned in the response's payload can be used as the `address` parameter in `GET /cryptocurrency/{address}`"
              }
            }
          },
          "400": {
            "description": "Input validation error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/definitions/Logic111bErrorRecipient"
                }
              }
            }
          },
          "503": {
            "description": "Sawtooth service unavailable",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/definitions/Sawtooth503Error"
                }
              }
            }
          }
        }
      }
    },
    "/cryptocurrency/{address}": {
      "get": {
        "tags": ["cryptocurrency"],
        "summary": "Retrieves one specific transaction",
        "description": "Retrieves one specific transaction by its signature. The transaction must exist in the blockchain, otherwise the endpoint's response will be a 400 error.",
        "operationId": "getTransactionByAddress",
        "parameters": [
          {
            "name": "address",
            "in": "path",
            "description": "The signature of the transaction to retrieve",
            "required": true,
            "schema": {
              "type": "string",
              "format": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Retrieves the specified transaction",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/definitions/NonExpandedTransaction"
                }
              }
            }
          },
          "503": {
            "description": "Sawtooth service unavailable",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/definitions/Sawtooth503Error"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": ["cryptocurrency"],
        "summary": "Updates and or approves or rejects a transaction",
        "description": "This endpoint allows to update non sensitive information of the transaction, such as its `description` **only** if the transaction has not be approved. Also, pending transactions can be approved or rejected using this endpoint.",
        "operationId": "modifyTransaction",
        "parameters": [
          {
            "name": "address",
            "in": "path",
            "description": "The signature of the transaction to update.",
            "required": true,
            "schema": {
              "type": "string",
              "format": "sha_256"
            }
          },
          {
            "in": "query",
            "name": "approve",
            "description": "Boolean representing whether the transaction should be approved or not",
            "required": false,
            "type": "boolean",
            "enum": ["true", "false"]
          }
        ],
        "responses": {
          "200": {
            "description": "Updated transaction",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/definitions/UpdatedTransaction"
                }
              }
            }
          },
          "503": {
            "description": "Sawtooth service unavailable",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/definitions/Sawtooth503Error"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "definitions": {
    "SimpleTransaction": {
      "type": "object",
      "required": ["amount", "sender"],
      "properties": {
        "sender": {
          "type": "number",
          "format": "double",
          "example": "jsbravoc"
        },
        "amount": {
          "type": "number",
          "format": "double",
          "example": "2.8"
        },
        "recipient": {
          "type": "string",
          "format": "string",
          "example": "jsbravoc2"
        },
        "pending": {
          "type": "boolean",
          "format": "boolean",
          "example": false
        },
        "signature": {
          "type": "string",
          "format": "string",
          "example": "0xc79eec8f62d09794feeb0677f2330d70cee5c640a03b2f35e508f8fe12140ff8754d37904ed443d942363ad3ebbd5f006536794f4110355343c0b91f4f2a48221b"
        },
        "valid_thru": {
          "type": "string",
          "format": "date-time"
        }
      }
    },
...
    "NonExpandedTransaction": {
      "type": "object",
      "required": ["sender", "amount", "recipient"],
      "properties": {
        "sender": {
          "type": "string",
          "format": "string",
          "example": "jsbravoc"
        },
        "amount": {
          "type": "number",
          "format": "double",
          "example": "2.8",
          "minimum": 0
        },
        "recipient": {
          "type": "string",
          "format": "string",
          "example": "jsbravoc2"
        },
        "valid_thru": {
          "type": "string",
          "format": "date-time"
        },
        "valid": {
          "type": "boolean",
          "format": "string",
          "example": "true"
        },
        "signature": {
          "type": "string",
          "format": "string",
          "example": "0xc79eec8fsa979saeeb8477f2330d70cee5c640a03b2f35e508dase12140ff8as4d37904ed443d942363ac3ebbd5f006536794f4110355553d08ff658ta5414fs"
        },
        "supporting_transactions": {
          "type": "array",
          "items": {
            "type": "string",
            "format": "string",
            "example": "0xc79eec8f62d09794feeb0677f2330d70cee5c640a03b2f35e508f8fe12140ff8754d37904ed443d942363ad3ebbd5f006536794f4110355343c0b91f4f2a48221b"
          }
        }
      }
    },
...
    "111bErrorRecipient": {
      "type": "object",
      "required": ["errorCode", "msg", "error"],
      "properties": {
        "errorCode": {
          "type": "string",
          "format": "string",
          "example": "111b"
        },
        "msg": {
          "type": "string",
          "format": "string",
          "example": "User with address created.user does not exist"
        },
        "error": {
          "type": "string",
          "format": "string",
          "example": "The user created.user does not exist"
        },
        "location": {
          "type": "string",
          "format": "string",
          "example": "body"
        },
        "param": {
          "type": "string",
          "format": "string",
          "example": "recipient"
        },
        "value": {
          "type": "string",
          "format": "string",
          "example": "created.user"
        }
      }
    },
    "503Error": {
      "type": "object",
      "required": ["errorCode", "msg", "error"],
      "properties": {
        "errorCode": {
          "type": "string",
          "format": "string",
          "example": "503"
        },
        "msg": {
          "type": "string",
          "format": "string",
          "example": "Hyperledger Sawtooth REST API is temporarily unavailable"
        },
        "error": {
          "type": "string",
          "format": "string",
          "example": "Hyperledger Sawtooth REST API is temporarily unavailable"
        }
      }
    },
...
    "UpdatedTransaction": {
      "type": "object",
      "properties": {
        "msg": {
          "type": "string",
          "format": "string",
          "example": "Transaction updated"
        },
        "payload": {
          "$ref": "#/definitions/NonExpandedTransaction"
        }
      }
    },
...
    "Sawtooth503Error": {
      "type": "object",
      "properties": {
        "errors": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/503Error"
          }
        }
      }
    },
...
    "Logic111bErrorRecipient": {
      "type": "object",
      "properties": {
        "errors": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/111bErrorRecipient"
          }
        }
      }
    },
...
  },
...
}